small tweaks to visual, removed scroll to pos, need to use ajax

+++ www/Models/Product.php
@@ -103,7 +103,7 @@ class Product extends SQL
             /**
      * @return float
      */
-    public function getPrice(): int
+    public function getPrice(): float
     {
         return $this->price;
     }
